+++ examples/example_com_module.py
@@ -23,9 +23,9 @@ from jsktoolbox.attribtool import ReadOnlyClass
 from jsktoolbox.raisetool import Raise
 from jsktoolbox.datetool import Timestamp
 
-from libs.base.classes import BModule
+from libs.base import ModuleMixin
 from libs.interfaces.modules import IComModule
-from libs.base.classes import BModuleConfig
+from libs.base import ModuleConfigMixin
 from libs.templates.modules import TemplateConfigItem
 from libs.com.message import Message, Multipart
 from libs.tools.datetool import MDateTime
@@ -39,11 +39,11 @@ class _Keys(object, metaclass=ReadOnlyClass):
     """
 
 
-class _ModuleConf(BModuleConfig):
+class _ModuleConf(ModuleConfigMixin):
     """Module Config private class."""
 
 
-class MExample(Thread, ThBaseObject, BModule, IComModule):
+class MExample(Thread, ThBaseObject, ModuleMixin, IComModule):
     """Example alert module."""
 
     def __init__(