+++ modules/run/memailtest.py
@@ -21,9 +21,9 @@ from jsktoolbox.attribtool import ReadOnlyClass
 from jsktoolbox.raisetool import Raise
 from jsktoolbox.datetool import Timestamp
 
-from libs.base.classes import BModule
+from libs.base import ModuleMixin
 from libs.interfaces.modules import IRunModule
-from libs.base.classes import BModuleConfig
+from libs.base import ModuleConfigMixin
 from libs.templates.modules import TemplateConfigItem
 from libs.com.message import Message, Multipart, Channel
 from libs.app import AppName
@@ -33,11 +33,11 @@ class _Keys(object, metaclass=ReadOnlyClass):
     """Define internal key names for the module."""
 
 
-class _ModuleConf(BModuleConfig):
+class _ModuleConf(ModuleConfigMixin):
     """Provide typed access to the module configuration."""
 
 
-class MEmailtest(Thread, ThBaseObject, BModule, IRunModule):
+class MEmailtest(Thread, ThBaseObject, ModuleMixin, IRunModule):
     """Generate sample multipart e-mail messages for communication tests."""
 
     def __init__(